Making INpmPackageInfoFactory + INpmPackageSearch public
