+++ src/LibraryManager/Properties/AssemblyInfo.cs
@@ -6,4 +6,3 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Microsoft.Web.LibraryManager.IntegrationTest")]
 [assembly: InternalsVisibleTo("Microsoft.Web.LibraryManager.Vsix")]
 [assembly: InternalsVisibleTo("Microsoft.Web.LibraryManager.Vsix.Test")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
