+++ src/LibraryManager/Providers/Unpkg/NpmPackageSearch.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
@@ -12,9 +11,7 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
 {
     internal sealed class NpmPackageSearch : INpmPackageSearch
     {
-        public const string NpmPackageInfoUrl = "https://registry.npmjs.org/{0}";
         private const string NpmPackageSearchUrl = "https://registry.npmjs.org/-/v1/search?text={0}&size=100"; // API doc at https://github.com/npm/registry/blob/master/docs/REGISTRY-API.md
-        public const string NpmLatestPackgeInfoUrl = "https://registry.npmjs.org/{0}/latest";
         public const string NpmsPackageSearchUrl = "https://api.npms.io/v2/search?q={1}+scope:{0}";
 
         public async Task<IEnumerable<string>> GetPackageNamesAsync(string searchTerm, CancellationToken cancellationToken)
@@ -195,107 +192,6 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
             }
 
             return packageNames;
-        }
-
-        public async Task<NpmPackageInfo> GetPackageInfoAsync(string packageName, CancellationToken cancellationToken)
-        {
-            NpmPackageInfo packageInfo = null;
-
-            if (packageName.StartsWith("@", StringComparison.Ordinal))
-            {
-                packageInfo = await GetPackageInfoForScopedPackageAsync(packageName, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                packageInfo = await GetPackageInfoForUnscopedPackageAsync(packageName, cancellationToken).ConfigureAwait(false);
-            }
-
-            return packageInfo;
-        }
-
-        private static async Task<NpmPackageInfo> GetPackageInfoForScopedPackageAsync(string packageName, CancellationToken cancellationToken)
-        {
-            Debug.Assert(packageName.StartsWith("@", StringComparison.Ordinal));
-            string searchName = "@" + HttpUtility.UrlEncode(packageName.Substring(1));
-            NpmPackageInfo packageInfo = null;
-
-            return await CreatePackageInfoAsync(searchName, packageInfo, cancellationToken);
-        }
-
-        private static async Task<NpmPackageInfo> GetPackageInfoForUnscopedPackageAsync(string packageName, CancellationToken cancellationToken)
-        {
-            NpmPackageInfo packageInfo = null;
-
-            try
-            {
-                string packageInfoUrl = string.Format(NpmLatestPackgeInfoUrl, packageName);
-                JObject packageInfoJSON = await WebRequestHandler.Instance.GetJsonObjectViaGetAsync(packageInfoUrl, cancellationToken).ConfigureAwait(false);
-
-                if (packageInfoJSON != null)
-                {
-                    packageInfo = NpmPackageInfo.Parse(packageInfoJSON);
-                }
-            }
-            catch (Exception)
-            {
-                packageInfo = new NpmPackageInfo(
-                    packageName,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable);
-            }
-
-            return await CreatePackageInfoAsync(packageName, packageInfo, cancellationToken);
-        }
-
-        private static async Task<NpmPackageInfo> CreatePackageInfoAsync(string packageName, NpmPackageInfo packageInfo, CancellationToken cancellationToken)
-        {
-            try
-            {
-                string packageInfoUrl = string.Format(NpmPackageInfoUrl, packageName);
-                JObject packageInfoJSON = await WebRequestHandler.Instance.GetJsonObjectViaGetAsync(packageInfoUrl, cancellationToken).ConfigureAwait(false);
-
-                if (packageInfoJSON != null && packageInfo == null)
-                {
-                    packageInfo = NpmPackageInfo.Parse(packageInfoJSON);
-                }
-
-                JObject versionsList = packageInfoJSON["versions"] as JObject;
-                if (versionsList != null)
-                {
-                    List<SemanticVersion> semanticVersions = versionsList.Properties().Select(p => SemanticVersion.Parse(p.Name)).ToList<SemanticVersion>();
-                    string latestVersion = semanticVersions.Max().OriginalText;
-                    IList<SemanticVersion> filteredSemanticVersions = FilterOldPrereleaseVersions(semanticVersions);
-
-                    packageInfo = new NpmPackageInfo(packageInfo.Name, packageInfo.Description, latestVersion, packageInfo.Author, packageInfo.Homepage, packageInfo.License, filteredSemanticVersions);
-                }
-            }
-            catch (Exception)
-            {
-                packageInfo = new NpmPackageInfo(
-                    packageName,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable,
-                    Resources.Text.LibraryDetail_Unavailable);
-            }
-
-            return packageInfo;
-        }
-
-        internal static IList<SemanticVersion> FilterOldPrereleaseVersions(List<SemanticVersion> semanticVersions)
-        {
-            List<SemanticVersion> filteredVersions = new List<SemanticVersion>();
-            List<SemanticVersion> releasedVersions = semanticVersions.Where(sv => sv.PrereleaseVersion == null).ToList<SemanticVersion>();
-            SemanticVersion latestReleaseVersion = releasedVersions.Max();
-
-            filteredVersions.AddRange(releasedVersions);
-            filteredVersions.AddRange(semanticVersions.Where(sv => sv.PrereleaseVersion != null && sv.CompareTo(latestReleaseVersion) > 0));
-
-            return filteredVersions;
         }
     }
 }
