+++ src/LibraryManager/Providers/Unpkg/UnpkgCatalog.cs
@@ -17,16 +17,16 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
         public const string CacheFileName = "cache.json";
         public const string LibraryFileListUrlFormat = "https://unpkg.com/{0}@{1}/?meta"; // e.g. https://unpkg.com/jquery@3.3.1/?meta
         public const string LatestLibraryVersonUrl = "https://unpkg.com/{0}/package.json"; // e.g. https://unpkg.com/jquery/package.json
-        private readonly INpmPackageInfoCache _packageInfoCache;
+        private readonly INpmPackageInfoFactory _packageInfoFactory;
         private readonly INpmPackageSearch _packageSearch;
         private readonly string _providerId;
         private readonly ILibraryNamingScheme _libraryNamingScheme;
         private readonly ILogger _logger;
         private readonly IWebRequestHandler _webRequestHandler;
 
-        public UnpkgCatalog(string providerId, ILibraryNamingScheme namingScheme, ILogger logger, IWebRequestHandler webRequestHandler, INpmPackageInfoCache packageInfoCache, INpmPackageSearch packageSearch)
+        public UnpkgCatalog(string providerId, ILibraryNamingScheme namingScheme, ILogger logger, IWebRequestHandler webRequestHandler, INpmPackageInfoFactory packageInfoFactory, INpmPackageSearch packageSearch)
         {
-            _packageInfoCache = packageInfoCache;
+            _packageInfoFactory = packageInfoFactory;
             _packageSearch = packageSearch;
             _providerId = providerId;
             _libraryNamingScheme = namingScheme;
@@ -192,7 +192,7 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
 
                     foreach (string packageName in packageNames)
                     {
-                        NpmPackageInfo packageInfo = await _packageInfoCache.GetPackageInfoAsync(packageName, CancellationToken.None);
+                        NpmPackageInfo packageInfo = await _packageInfoFactory.GetPackageInfoAsync(packageName, CancellationToken.None);
 
                         var completionItem = new CompletionItem
                         {
@@ -212,7 +212,7 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
                     completionSet.Start = name.Length + 1;
                     completionSet.Length = version.Length;
 
-                    NpmPackageInfo npmPackageInfo = await _packageInfoCache.GetPackageInfoAsync(name, CancellationToken.None);
+                    NpmPackageInfo npmPackageInfo = await _packageInfoFactory.GetPackageInfoAsync(name, CancellationToken.None);
 
                     IList<SemanticVersion> versions = npmPackageInfo.Versions;
 
@@ -249,7 +249,7 @@ namespace Microsoft.Web.LibraryManager.Providers.Unpkg
             try
             {
                 IEnumerable<string> packageNames = await _packageSearch.GetPackageNamesAsync(term, CancellationToken.None);
-                libraryGroups = packageNames.Select(packageName => new UnpkgLibraryGroup(_packageInfoCache, packageName)).ToList<ILibraryGroup>();
+                libraryGroups = packageNames.Select(packageName => new UnpkgLibraryGroup(_packageInfoFactory, packageName)).ToList<ILibraryGroup>();
             }
             catch (Exception ex)
             {
