+++ test/LibraryManager.Test/Providers/Unpkg/UnpkgCatalogTest.cs
@@ -20,7 +20,7 @@ namespace Microsoft.Web.LibraryManager.Test.Providers.Unpkg
         private UnpkgCatalog SetupCatalog(IWebRequestHandler handler = null)
         {
             var packageSearch = new NpmPackageSearch();
-            var infoCache = new NpmPackageInfoCache(packageSearch);
+            var infoCache = new NpmPackageInfoFactory();
             return new UnpkgCatalog(UnpkgProvider.IdText,
                                     new VersionedLibraryNamingScheme(),
                                     new Logger(),
@@ -277,22 +277,21 @@ namespace Microsoft.Web.LibraryManager.Test.Providers.Unpkg
         {
             //Arrange
             var packageSearch = new Mock<INpmPackageSearch>();
+            var infoFactory = new Mock<INpmPackageInfoFactory>();
 
             var packages = new List<string>() { "testPkg" };
             packageSearch.Setup(p => p.GetPackageNamesAsync(It.Is<string>(s => string.Equals(s, "testPkg")), It.IsAny<CancellationToken>()))
                 .Returns(Task.FromResult((IEnumerable<string>)packages));
 
             var testPkgInfo = new NpmPackageInfo(name: "testPkg", description: "description", latestVersion: "1.2.3", author: "brzh", homepage: "https://testPkg.com", license: "MIT");
-            packageSearch.Setup(p => p.GetPackageInfoAsync(It.Is<string>(s => string.Equals(s, "testPkg")), It.IsAny<CancellationToken>()))
+            infoFactory.Setup(p => p.GetPackageInfoAsync(It.Is<string>(s => string.Equals(s, "testPkg")), It.IsAny<CancellationToken>()))
                 .Returns(Task.FromResult(testPkgInfo));
-
-            var infoCache = new NpmPackageInfoCache(packageSearch.Object);
 
             var sut = new UnpkgCatalog(UnpkgProvider.IdText,
                             new VersionedLibraryNamingScheme(),
                             new Logger(),
                             new Mocks.WebRequestHandler(),
-                            infoCache,
+                            infoFactory.Object,
                             packageSearch.Object);
 
             //Act
